+++ modules/App/assets/vue-components/fields/richtext/components/color.js
@@ -1,0 +1,44 @@
+let instanceCount = 0;
+
+export default {
+    data() {
+
+        return {
+            uid: `richtext-action-color-${++instanceCount}`
+        }
+    },
+
+    props: {
+        editor: {
+            type: Object,
+            default: null
+        }
+    },
+
+    computed: {
+        color() {
+            return this.editor.getAttributes('textStyle').color || '';
+        }
+    },
+
+    mounted() {
+
+    },
+
+    methods: {
+        pick() {
+            VueView.ui.modal('app:assets/vue-components/fields/richtext/dialogs/color.js', {color: this.color}, {
+
+                save: color => {
+                    this.editor.chain().focus().setColor(color).run();
+                }
+            }, {size: 'medium'});
+        }
+    },
+
+    template: /*html*/`
+        <div>
+            <icon @click="pick()" :style="{color}">format_color_text</icon>
+        </div>
+    `
+}